Move code out of try-except blocks
Other general cleanups. No functional change.
